Add multi-containers and special stackable parameters

+++ packager-api/src/main/java/com/developsoft/comtainer/rest/dto/ConfigDto.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter @Setter
 public class ConfigDto {
 	private Integer cargoSupport; //in percents
-	private Double lightUnstackableWeightLimit;	//derived from F)	 2nd  sub property for unstackable goods- to determine light goods that unstackable. 
+	private Float lightUnstackableWeightLimit;	//derived from F)	 2nd  sub property for unstackable goods- to determine light goods that unstackable. 
 	private boolean openNewContainer; //if Cargo does not fit in current containers
 
 }
